Amélioration de la gestion des permissions pour les opérations des utilisateurs

+++ src/main/java/com/example/demo/controller/ClientController.java
@@ -4,9 +4,13 @@ package com.example.demo.controller;
 import com.example.demo.dto.ClientDTO;
 import com.example.demo.dto.UserDto;
 import com.example.demo.entity.Client;
+import com.example.demo.enums.Role;
 import com.example.demo.exception.ApiResponse;
 import com.example.demo.service.ClientService;
+import com.example.demo.service.HelperService;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,9 +22,21 @@ import java.util.Map;
 public class ClientController {
 
     private final ClientService clientService;
+    private final HelperService helperService ;
 
     @PostMapping
-    public ClientDTO createClient(@RequestBody Map<String, Object> request) {
+    public ResponseEntity<?> createClient(
+            @RequestBody Map<String, Object> request,
+            HttpServletRequest httpRequest) {
+
+        if(!helperService.isAuthenticated(httpRequest)) {
+            return ResponseEntity.status(401).body("tu ne pas login");
+        }
+
+        if (!helperService.hasRole(httpRequest, Role.ADMIN)) {
+            return ResponseEntity.status(403).body("vous n'avez pas la permission pour faire cette action : ADMIN only");
+        }
+
         Map<String, Object> clientData = (Map<String, Object>) request.get("client");
         Map<String, Object> userData = (Map<String, Object>) request.get("user");
 
@@ -29,21 +45,21 @@ public class ClientController {
         clientDTO.setEmail((String) clientData.get("email"));
         clientDTO.setLoyaltyLevel(Enum.valueOf(
                 com.example.demo.enums.LoyaltyLevel.class,
-                (String) clientData.get("loyaltyLevel"))
-        );
+                (String) clientData.get("loyaltyLevel")
+        ));
         clientDTO.setTotalOrders((Integer) clientData.get("totalOrders"));
         clientDTO.setTotalSpent((Double) clientData.get("totalSpent"));
 
         UserDto userDto = new UserDto();
         userDto.setUsername((String) userData.get("username"));
-        userDto.setPassword((String) userData.get("password"));
         userDto.setRole(Enum.valueOf(
                 com.example.demo.enums.Role.class,
-                (String) userData.get("role"))
-        );
+                (String) userData.get("role")
+        ));
 
-        return clientService.create(clientDTO, userDto);
+        return ResponseEntity.ok(clientService.create(clientDTO, userDto));
     }
+
 
     @GetMapping
     public List<ClientDTO> getAll()
@@ -52,8 +68,19 @@ public class ClientController {
     }
 
     @PostMapping("/update/{id}")
-    public ClientDTO updateClient(@PathVariable Long id, @RequestBody Map<String, Object> request)
-    {
+    public ResponseEntity<?> updateClient(
+            @PathVariable Long id,
+            @RequestBody Map<String, Object> request,
+            HttpServletRequest httpRequest) {
+
+        if (!helperService.isAuthenticated(httpRequest)) {
+            return ResponseEntity.status(401).body("tu ne pas login");
+        }
+
+        if (!helperService.hasRole(httpRequest, Role.ADMIN)) {
+            return ResponseEntity.status(403).body("vous n'avez pas la permission pour faire cette action : ADMIN only");
+        }
+
         Map<String, Object> clientData = (Map<String, Object>) request.get("client");
         Map<String, Object> userData = (Map<String, Object>) request.get("user");
 
@@ -70,26 +97,43 @@ public class ClientController {
 
         UserDto userDto = new UserDto();
         userDto.setUsername((String) userData.get("username"));
-        userDto.setPassword((String) userData.get("password"));
+        //userDto.setPassword((String) userData.get("password"));
         userDto.setRole(Enum.valueOf(
                 com.example.demo.enums.Role.class,
                 (String) userData.get("role"))
         );
 
-        return  clientService.update(id , clientDTO , userDto) ;
+        return ResponseEntity.ok(clientService.update(id, clientDTO, userDto));
 
     }
 
     @DeleteMapping("/{id}")
-    public ApiResponse deleteClient(@PathVariable Long id)
+    public ResponseEntity<?> deleteClient(@PathVariable Long id , HttpServletRequest httpRequest)
     {
-        return clientService.delete(id);
+        if (!helperService.isAuthenticated(httpRequest)) {
+            return ResponseEntity.status(401).body("tu ne pas login");
+        }
+
+        if (!helperService.hasRole(httpRequest, Role.ADMIN)) {
+            return ResponseEntity.status(403).body("vous n'avez pas la permission pour faire cette action : ADMIN only");
+        }
+
+        ApiResponse response = clientService.delete(id);
+        return ResponseEntity.ok(response);
     }
 
     @GetMapping("/{id}")
-    public ClientDTO getById(@PathVariable Long id)
+    public ResponseEntity<?> getById(@PathVariable Long id , HttpServletRequest httpRequest)
     {
-        return clientService.getById(id) ;
+        if (!helperService.isAuthenticated(httpRequest)) {
+            return ResponseEntity.status(401).body("tu ne pas login");
+        }
+
+        if (!helperService.hasRole(httpRequest, Role.ADMIN)) {
+            return ResponseEntity.status(403).body("vous n'avez pas la permission pour faire cette action : ADMIN only");
+        }
+        ClientDTO response =  clientService.getById(id) ;
+        return   ResponseEntity.ok(response) ;
     }
